feat: Enhanced supply graph API integration with port 8081
- Updated supply graph API to use port 8081 instead of 8001
- Enhanced OKH data fetching with multiple endpoint support
- Improved error handling and fallback mechanisms
- Added comprehensive logging for debugging
- Enhanced UI with better visual representation of supply trees
- Renamed test-supply-graph-api.vue to supply-graph-api.vue
- Added Azure Blob Storage configuration support
- Integrated project data platform with supply graph AI service
